requirements.txt
openai
pandas
pypdf
python-dotenv
requests
google-generativeai